+++ src/background/global.js
@@ -79,7 +79,7 @@ const Global = {
   getFilename: (path) => Global.getBasename(path).replace(/\?.*$/, "").replace(/:.*$/, ""), // strip twitter-style tags "?format=jpg" and/or ":large"
 
   // sanitize filename, remove extension
-  getFilePart: (path) => Global.getFilename(path).replace(/\.[^.]*$/, ""), // strip extension
+  getFilePart: (path) => Global.getFilename(path).replace(/\.[^.]+$/, ""), // strip extension
 
   getFileExt: (path) => {
     const m = Global.getFilename(path).match(/\.([^./]+)$/);
@@ -97,6 +97,7 @@ const Global = {
   // replace all invalid characters and slashes
   sanitizeFilename: (filename, str = "_") =>
     filename
+      .replace(/\u200b/g, "") // remove zero-width spaces
       .replace(/[*"/\\:<>|?]/g, str) // replace invalid characters
       .replace(/^[\s]+/, "") // strip leading spaces
       .replace(/[.\s]+$/, ""), // strip trailing spaces and period
@@ -104,6 +105,7 @@ const Global = {
   // replace invalid characters and strip leading/trailing slashes
   sanitizePath: (path, str = "_") =>
     path
+      .replace(/\u200b/g, "") // remove zero-width spaces
       .replace(/[*":<>|?]/g, str) // replace invalid characters
       .replace(/[/\\]+/g, "/") // replace backslash with forward slash
       .replace(/^[/]/, "") // strip leading slash